Refactor color handling to DrawColor type
Change DrawColor to use uint8_t channels and update its constructor. Replace multiple sprite drawing APIs to accept a DrawColor parameter (glBlitSquare, glDrawBox, etc.) and update call sites accordingly. Misc cleanup: use auto in item lambdas, remove an unused iostream include, simplify creature frame resolution, and tidy duplicated includes in brush overlay. These changes centralize color representation and reduce repetitive argument lists when passing colors.

+++ source/app/definitions.h
@@ -154,12 +154,12 @@ constexpr int PixelFormatRGB = 3;
 constexpr int PixelFormatRGBA = 4;
 
 struct DrawColor {
-	int r;
-	int g;
-	int b;
-	int a;
+	uint8_t r;
+	uint8_t g;
+	uint8_t b;
+	uint8_t a;
 
-	constexpr DrawColor(int red = 255, int green = 255, int blue = 255, int alpha = 255) :
+	constexpr DrawColor(uint8_t red = 255, uint8_t green = 255, uint8_t blue = 255, uint8_t alpha = 255) :
 		r(red),
 		g(green),
 		b(blue),
